refactor: 修复 P2/P3 代码质量问题
- P2: 空 catch 块添加 debug 日志（connection.ts）
- P2: 确认工具函数已通过 re-export 统一，无需改动
- P3: 统一 originalToolName 命名为 originalName（manager.ts + 测试）
- P3: 拆分 CustomMCP 管理为独立模块（custom-mcp.ts），委托模式保持 API 兼容
